Moved the Dataset object into the objects folder
