+++ vip_hci/objects/__init__.py
@@ -21,3 +21,4 @@ from .pploci import *
 from .ppmediansub import *
 from .ppnmf import *
 from .pppca import *
+from .dataset import *
